# src/main/generated/src/main/kotlin/com/example/usj_backend/api/MatchingApi.kt
/**
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.10.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
*/
package com.example.usj_backend.api

import com.example.usj_backend.model.LikeRequest
import com.example.usj_backend.model.LikeResponse
import com.example.usj_backend.model.Match
import com.example.usj_backend.model.MatchingCandidate
import io.swagger.v3.oas.annotations.*
import io.swagger.v3.oas.annotations.enums.*
import io.swagger.v3.oas.annotations.media.*
import io.swagger.v3.oas.annotations.responses.*
import io.swagger.v3.oas.annotations.security.*
import org.springframework.http.HttpStatus
import org.springframework.http.MediaType
import org.springframework.http.ResponseEntity

import org.springframework.web.bind.annotation.*
import org.springframework.validation.annotation.Validated
import org.springframework.web.context.request.NativeWebRequest
import org.springframework.beans.factory.annotation.Autowired

import jakarta.validation.constraints.DecimalMax
import jakarta.validation.constraints.DecimalMin
import jakarta.validation.constraints.Email
import jakarta.validation.constraints.Max
import jakarta.validation.constraints.Min
import jakarta.validation.constraints.NotNull
import jakarta.validation.constraints.Pattern
import jakarta.validation.constraints.Size
import jakarta.validation.Valid

import kotlin.collections.List
import kotlin.collections.Map

@RestController
@Validated
interface MatchingApi {

    @Operation(
        tags = ["matching",],
        summary = "マッチング候補取得",
        operationId = "apiMatchingCandidatesGet",
        description = """現在のユーザーに対するマッチング候補を取得します""",
        responses = [
            ApiResponse(responseCode = "200", description = "マッチング候補取得成功", content = [Content(array = ArraySchema(schema = Schema(implementation = MatchingCandidate::class)))]),
            ApiResponse(responseCode = "401", description = "未認証")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/api/matching/candidates"],
            produces = ["application/json"]
    )
    fun apiMatchingCandidatesGet(@Min(1) @Max(50) @Parameter(description = "", schema = Schema(defaultValue = "10")) @Valid @RequestParam(value = "limit", required = false, defaultValue = "10") limit: kotlin.Int): ResponseEntity<List<MatchingCandidate>>

    @Operation(
        tags = ["matching",],
        summary = "いいね送信",
        operationId = "apiMatchingLikePost",
        description = """指定したユーザーにいいねを送信します""",
        responses = [
            ApiResponse(responseCode = "200", description = "いいね送信成功", content = [Content(schema = Schema(implementation = LikeResponse::class))]),
            ApiResponse(responseCode = "400", description = "入力エラー"),
            ApiResponse(responseCode = "401", description = "未認証")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.POST],
            value = ["/api/matching/like"],
            produces = ["application/json"],
            consumes = ["application/json"]
    )
    fun apiMatchingLikePost(@Parameter(description = "", required = true) @Valid @RequestBody likeRequest: LikeRequest): ResponseEntity<LikeResponse>

    @Operation(
        tags = ["matching",],
        summary = "マッチ一覧取得",
        operationId = "apiMatchingMatchesGet",
        description = """成立したマッチの一覧を取得します""",
        responses = [
            ApiResponse(responseCode = "200", description = "マッチ一覧取得成功", content = [Content(array = ArraySchema(schema = Schema(implementation = Match::class)))]),
            ApiResponse(responseCode = "401", description = "未認証")
        ],
        security = [ SecurityRequirement(name = "basicAuth"),SecurityRequirement(name = "bearerAuth") ]
    )
    @RequestMapping(
            method = [RequestMethod.GET],
            value = ["/api/matching/matches"],
            produces = ["application/json"]
    )
    fun apiMatchingMatchesGet(): ResponseEntity<List<Match>>
}